clean up unused configs and constants (#58)
* clean up unused configs and constants

* Add back DataPerDPR

+++ pallets/credit/src/mock.rs
@@ -127,7 +127,7 @@ parameter_types! {
     pub const CreditCapTwoEras: u8 = CREDIT_CAP_TWO_ERAS;
     pub const CreditAttenuationStep: u64 = CREDIT_ATTENUATION_STEP;
     pub const MinCreditToDelegate: u64 = 100;
-    pub const MicropaymentToCreditFactor: u64 = 1_000_000_000_000_000;
+    pub const MicropaymentToCreditFactor: u128 = 1_000_000_000_000_000;
     pub const BlocksPerEra: BlockNumber =  6 * EPOCH_DURATION_IN_BLOCKS;
 }
